Remove client-go uses
Controllers are supposed to use the client given to them by the
manager, instead of using client-go. This removes our uses of
client-go, except (for now) in tests.

Ultimately this will simplify the operator rebuild, since we won't
have to take care of rebuilding the clientset.

The clientset itself can't be removed yet, subctl imports it.

Signed-off-by: Stephen Kitt <skitt@redhat.com>